[BACKEND] Fix an issue with the pipeliner (#4247)
During pipelining operations that do not depend on or being dependent by
anchor operations are considered remaining ops and will be scheduled
into the last stage. These ops do not present in the existing stages but
can be visited by other staged ops. Fixing an ICE when looking for
clusters for them.