[BE][PIPELINE] Add fix for the wgmma pipelining bug with subview dist 1 (#4579)
If the subview for the mma operand is in the previous iteration of the
loop, it would be skipped by the code that looks for it, if it feeds
directly into the loop yield. This PR fixes that issue allowing for them
subview to be found and enabling the pipelining in some cases skipped
previously.