[AMD] Introduce a CanonicalizePointer pass (#4241)
In order to support buffer operations (see Chapter 9 of the [MI300 ISA
reference](https://www.amd.com/content/dam/amd/en/documents/instinct-tech-docs/instruction-set-architectures/amd-instinct-mi300-cdna3-instruction-set-architecture.pdf))
the pointers we feed to the `buffer_load`/`buffer_store` instructions
need to be constant across the threads.

This means that they have to point at some base address. The only thing
that can vary per-thread is the offset.

If the base address is divergent (i.e., it is different per each thread)
the LLVM backend will emit a waterfall loop (i.e., `for i=0:64: mask all
the threads except the i-th; buffer_load for thread i`)

This pass is canonicalizing the pointers across the IR, in this
<basePtr,offset> style.

- Propagate an offset across the IR associated with the given pointer.
For each operation that makes use of a pointer we replace the pointer
with the `basePtr` and we add an `offset` operand that tracks where the
pointer is.
- We replace `tt.addptr %ptr, %offset` with `%newOffset = arith.add
%base_offset, %offset` and propagate `newOffset`
- We materialize `tt.addptr` only before a `tt.load` or a `tt.store`
operation.

This PR only adds this canonicalization. A subsequent PR will run this
pass just before translating TritonIR into `LLVMIR` (and if the pass
succeeds, we will enable buffer operations).

**Note**: enabling buffer operations will give us a way to support
masking without having to emit branches.

## Correctness issues 
It's worth underlying that when we carry out the canonicalization, we
are (by default) accumulating into a 64 bit offset. This is because the
following IR
```
%ptr0 = addptr %ptr, %x
%ptr1 = addptr %ptr0, %y
%ptr2 = addptr %ptr1, %z
```
Is not, in general, equivalent to this following one:
```
%sum_32bits = %x + %y + %z
%fatPtr = addptr %ptr, %sum
```
Because `%x+%y+%z` might overflow. 
However, there are shortcut we can take to circumvent this problem.
a) If the offset is constant across the program (i.e., does not come
from a `make_range`) we add this scalar into the scalar pointer
b) if the non-constant offset is added only once, and if it is 32 bits,
we can simply use its 32-bit value. I.e.,
```
%ptr0 = addptr %ptr, %x
```
Here `%x` can stay 32 bits, we don't need to extend to 64.

There is a third shortcut that we will investigate even further: it
looks like Triton IR is written assuming that 32 bit addition cannot
overflow. If this is the case then we can assume that `sum_32bits` does
not overflow and always simplify the IR when the offsets are 32-bits
values.

---------

Co-authored-by: Lei Zhang <antiagainst@gmail.com>