[AMD] Remove stream pipeliner v1 (#4845)
We have flipped stream pipeliner v2 on as default
for quite sometime. All known issues has been fixed.
So now remove old v1 pipeliner.

Note that this changes know `num_stages` are handled:
previously we used to enable pipelining if `num_stages`
is `0`, which really is not a good behavior. Now switched
to follow common practice where `0`/`1` won't trigger
pipelining anymore; need `2` or more to trigger. 

Given downstream users might be using `0` in the 
codebase, right now we `assert` to give developers
a clear indication the switch of behavior instead of
silently drop the perf. The `assert` is expected to be
dropped sometime down the line.

---------

Co-authored-by: Lei Zhang <antiagainst@gmail.com>